requirements.txt
fastapi[standard]==0.109.2
pydantic-settings==2.2.1
sqlmodel==0.0.16
uvicorn==0.27.1
asyncpg==0.29.0

###
black
isort
flake8
alembic==1.13.1
